further refactoring and decoupling
