+++ project/src/main/java/com/shurik/droidzebra/InvalidMove.java
@@ -1,0 +1,8 @@
+package com.shurik.droidzebra;
+
+/**
+ * Created by stefan on 18.03.2018.
+ */
+public class InvalidMove extends Exception {
+    private static final long serialVersionUID = 8970579827351672330L;
+}
